+++ src/main/java/com/yanny/ytech/configuration/block_entity/AqueductValveBlockEntity.java
@@ -14,6 +14,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
+import net.minecraft.world.level.material.FlowingFluid;
 import net.minecraft.world.level.material.FluidState;
 import net.minecraft.world.level.material.Fluids;
 import org.jetbrains.annotations.NotNull;
@@ -84,7 +85,8 @@ public class AqueductValveBlockEntity extends IrrigationBlockEntity {
         if (flow > 0) {
             IrrigationServerNetwork network = YTechMod.IRRIGATION_PROPAGATOR.server().getNetwork(this);
 
-            if (network != null && network.getFluidHandler().getFluidAmount() < network.getFluidHandler().getCapacity()) {
+            if (network != null && level.random.nextInt(YTechMod.CONFIGURATION.getValveChanceToDrainWater()) == 0
+                    && network.getFluidHandler().getFluidAmount() + YTechMod.CONFIGURATION.getValveFillAmount() <= network.getFluidHandler().getCapacity()) {
                 Set<BlockPos> checkedBlocks = new HashSet<>();
 
                 for (BlockPos pos : getValidNeighbors()) {
@@ -119,7 +121,13 @@ public class AqueductValveBlockEntity extends IrrigationBlockEntity {
     private int calculateFlow(@NotNull ServerLevel level) {
         return getValidNeighbors().stream().anyMatch((pos) -> {
             BlockState blockState = level.getBlockState(pos);
-            return blockState.getBlock() == Blocks.WATER;
+
+            if (blockState.getBlock() == Blocks.WATER) {
+                FluidState fluidState = blockState.getFluidState();
+                return !fluidState.hasProperty(FlowingFluid.FALLING) || !fluidState.getValue(FlowingFluid.FALLING);
+            }
+
+            return false;
         }) ? YTechMod.CONFIGURATION.getValveFillAmount() : 0;
     }